requirements.txt
anyio==4.9.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
